bind::tools: initial implementation
Add initial support for installing the BIND tools.  The package names
vary from OS to OS, and there's the additional complication that the
bind-tools package on some OSes conflicts with the bind-server
package, so the bind::tools class has to be told whether or not the
server is being installed.  For this reason, the bind::tools class
should be considered an internal implementation detail of the bind
class.

Right now the bind class can't do anything *but* install the tools.