implement view configuration templating
This implements the initial templating for view configuration, and
also adds the ability for views to specify defaults inherited by all
zones in two different ways: either configured at the view level (for
when that makes sense and BIND supports it -- e.g., allow-transfer) or
explicitly configured for each zone in the view (when the zone
configuration doesn't explicitly override it).  Note that zones that
are manually instantiated as bind::zone objects will inherit the first
kind of defaults but not the second kind, because the defaulting is
done by the create_resources() library function -- but you can
presumably do this in hiera data as well.  (Does that make too many
different ways of specifying the same behavior? Probably.)

For certain configurable parameter, if the option value is not defined
(or is defined to be the undefined value), it will be omitted from the
configuration, giving you whatever BIND's default is.