Implement TSIG support
This includes support for configuring TSIG keys (unfortunately not
generating them), and support for generating the "server" statements
in the BIND configuration file that associate a key with an IP
address.  No support for ACLs yet (but that's coming), which means no
master-side support and no secure NOTIFY.  It would be really great if
there were some glue for

  (a) automatically generating random keys for each pair of servers, and

  (b) automatically configuring the right server statements to
      associate each key with the known-to-puppet IP addresses of
      the corresponding servers.

This is left for future work, if it ever bugs someone enough to
implement.