bind::config: ensure that the directories specified actually exist
This requires actually knowing what the correct ownership and mode is
for these directories, which unfortunately varies from OS to OS.  For
now, there's a single switch for "Debian-like" (root:bind 0775) versus
"FreeBSD-like" (bind:bind 0755) directory permissions.  The goal here
is that directories that we require, if included in the BIND package,
should exactly match the package, and those that aren't in the package
should match an analogous package-provided directory when possible.