bind::zone::slave: implement "bump in the wire" signing
Sometimes it's useful -- for transition purposes or because of how
your provisioning system is built -- to do DNSsec signing on a slave
server.  BIND 9.9's "inline signing" function supports this, so we
should allow the same inline_signing parameter on bind::zone::slave as
we do on bind::zone::master.