data/common.yaml: explicitly specify tcp-listen-queue
For now, we are doing this by changing the default options, but
this should be hoisted into a separate class parameter for
bind::config.

With modern changes to EDNS0 buffer sizes and DNSsec, clients fall
back to TCP far more often than before.  The default listen queue size
of 16 is inadequate; bump it up to 128 (which would be the default on
FreeBSD if named didn't override it).