bind::config: use modern argument type-checking style, add dnssec_policies
Convert all arguments to modern Puppet typed arguments.

Add bind::config::dnssec_policies which is a hash that will
instantiate a policy in the config for every key, which must be in the
form of parameters for the bind::dnssec_policy type.  No syntax
checking is done on the policy name itself.